feat(scm-multi-platform-detetion): Exporting shared code to a new module
